refactor(driver): extract block-build + forkchoice methods to driver/build.rs (refactor 2.1h)
Third impl-block extraction of Phase 2.1. Moves the three methods
that bridge driver state and the reth consensus engine API.

- `crates/based-rollup/src/driver/build.rs` (new, 170 lines):
  - `build_and_insert_block` — build the next sequential block via
    `build_derived_block`, submit via `engine.new_payload`, promote
    via `update_fork_choice`. The canonical entry point used by
    every step_* function to advance the L2 head.
  - `fork_choice_updated_with_retry` — exponential backoff retry
    on transient SYNCING responses from the engine.
  - `update_fork_choice` — happy-path FCU that mutates driver state
    only AFTER the engine confirms VALID (invariant: engine/driver
    must never desync).
  - `build_derived_block` stays in mod.rs for now — it's the
    payload-building half with cross-references to the still-in-mod
    protocol-tx path (split deferred to 2.1i).

- `crates/based-rollup/src/driver/mod.rs` (-110 lines → 4134 total):
  - `mod build;` declaration.
  - Three method bodies removed from the main `impl Driver` block.
  - Obsolete imports pruned: `ForkchoiceState`, `ForkchoiceUpdated`,
    `PayloadAttributes`, `EngineApiMessageVersion` (all now live
    in build.rs only).
  - `FCU_SYNCING_INITIAL_BACKOFF_MS`, `FCU_SYNCING_MAX_RETRIES`,
    `compute_forkchoice_state` moved from the unconditional
    `use types::{…}` list to a `#[cfg(test)]`-gated use. The driver
    tests file (`driver_tests.rs`) still references them directly
    via `use super::*`, so they need to be in scope in test builds;
    release builds no longer import them (they're only used from
    build.rs and rewind.rs which import from `super::types::`
    directly).

Cumulative progress:
- mod.rs: 5509 → 4134 lines (-1375 / -25.0%) — **past the quarter mark**.
- 7 sibling modules totaling 1622 lines.

`cargo check`, `cargo clippy --all-features -- -D warnings`, and
`cargo nextest -p based-rollup` (608/608) all pass. Zero behavioral
change.

Next: 2.1i — verify.rs (the biggest extraction by LOC, ~750 lines
including `classify_and_apply_verification`, the four filtering
methods, trial_execute, compute_intermediate_roots).
